Remove use of legacy FIL from Python cuML (#6728)
Part of #6506.

This PR completes the migration from legacy FIL to the new experimental FIL implementation across cuML's Random Forest estimators and related components.

It does not remove all legacy FIL C++ components.

## Major Changes

### FIL Migration
* **Remove legacy FIL**: Completely removes `cuml/legacy/fil` module
* **Update cuML RF estimators**: Migrate RandomForest classifier and regressor to use new FIL for GPU prediction
* **Simplify serialization**: Replace custom serialization with Treelite's native `serialize_bytes()` and `deserialize_bytes()` methods
* **New FIL interface**: Update parameter names from deprecated `algo`/`fil_sparse_format` to new `layout`/`default_chunk_size`/`align_bytes`

### API Changes & Deprecations
* **Parameter deprecation**: Add deprecation warnings for `algo`, `fil_sparse_format`, and `output_class` parameters (removal in 25.08)
* **New parameters**: Introduce `layout`, `default_chunk_size`, `align_bytes`, and `is_classifier` parameters
* **Method updates**: Replace `to_treelite_checkpoint()` with `serialize()` in documentation and tests
* **Compatibility layer**: Maintain backwards compatibility through compatibility layer.

### Dask Integration Updates
* **Remove float64 restriction**: Allow float64 data in Dask RF (previously restricted to float32)
* **Update concatenation**: Use Treelite's native concatenation instead of custom implementation
* **Simplify model handling**: Replace handle-based operations with direct Treelite model objects

### Test Infrastructure
* **Comprehensive test migration**: Update all Random Forest and FIL tests to use new API
* **Backward compatibility tests**: Create `test_random_forest_compat.py` and `test_fil_compat.py` to explicitly test backwards-compatibility (to be removed in 25.08)
* **Parameter validation**: Capture warnings for deprecated parameters in tests
* **Remove experimental tests**: Migrate tests from `test_filex.py` to main `test_fil.py`

### Documentation & Examples
* **Update notebooks**: Fix Treelite serialization examples in documentation
* **API documentation**: Add comprehensive docstrings for new FIL interface

### Build System
* **Remove legacy build**: Remove `cuml/legacy/fil/CMakeLists.txt` and associated build configuration
* **Add Treelite support**: Update `cuml/internals/CMakeLists.txt` to link Treelite libraries
* **Clean dependencies**: Remove legacy FIL from main CMakeLists.txt

Authors:
  - Philip Hyunsu Cho (https://github.com/hcho3)
  - Simon Adorf (https://github.com/csadorf)

Approvers:
  - Divye Gala (https://github.com/divyegala)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6728
