+++ python/cuml/CMakeLists.txt
@@ -101,7 +101,6 @@ add_subdirectory(cuml/decomposition)
 add_subdirectory(cuml/ensemble)
 add_subdirectory(cuml/explainer)
 add_subdirectory(cuml/experimental/fil)
-add_subdirectory(cuml/legacy/fil)
 add_subdirectory(cuml/kernel_ridge)
 add_subdirectory(cuml/linear_model)
 add_subdirectory(cuml/manifold)
