+++ python/cuml/cuml/accel/tests/scikit-learn/xfail-list.yaml
@@ -485,7 +485,6 @@
   - "sklearn.ensemble.tests.test_voting::test_voting_verbose[estimator0]"
   - "sklearn.ensemble.tests.test_voting::test_voting_verbose[estimator1]"
   - "sklearn.ensemble.tests.test_voting::test_none_estimator_with_weights[X0-y0-voter0]"
-  - "sklearn.ensemble.tests.test_voting::test_none_estimator_with_weights[X1-y1-voter1]"
   - "sklearn.ensemble.tests.test_voting::test_predict_on_toy_problem[42]"
   - "sklearn.ensemble.tests.test_voting::test_sample_weight[42]"
   - "sklearn.ensemble.tests.test_voting::test_set_estimator_drop"
@@ -500,12 +499,7 @@
   - "sklearn.feature_selection.tests.test_rfe::test_rfe_features_importance"
   - "sklearn.inspection.tests.test_partial_dependence::test_multiclass_multioutput[KNeighborsClassifier]"
   - "sklearn.inspection.tests.test_partial_dependence::test_multiclass_multioutput[RandomForestClassifier]"
-  - "sklearn.inspection.tests.test_partial_dependence::test_partial_dependence_sample_weight_with_recursion"
   - "sklearn.inspection.tests.test_partial_dependence::test_recursion_decision_tree_vs_forest_and_gbdt[0]"
-  - "sklearn.inspection.tests.test_permutation_importance::test_permutation_importance_correlated_feature_regression[ones-0.5-1]"
-  - "sklearn.inspection.tests.test_permutation_importance::test_permutation_importance_correlated_feature_regression[ones-0.5-2]"
-  - "sklearn.inspection.tests.test_permutation_importance::test_permutation_importance_correlated_feature_regression[ones-1.0-1]"
-  - "sklearn.inspection.tests.test_permutation_importance::test_permutation_importance_correlated_feature_regression[ones-1.0-2]"
   - "sklearn.inspection.tests.test_permutation_importance::test_robustness_to_high_cardinality_noisy_feature[0.5-1]"
   - "sklearn.inspection.tests.test_permutation_importance::test_robustness_to_high_cardinality_noisy_feature[0.5-2]"
   - "sklearn.inspection.tests.test_permutation_importance::test_robustness_to_high_cardinality_noisy_feature[1.0-1]"
