+++ python/cuml/cuml/ensemble/__init__.py
@@ -15,9 +15,5 @@
 #
 
 
-from cuml.ensemble.randomforest_common import (
-    _check_fil_parameter_validity,
-    _obtain_fil_model,
-)
 from cuml.ensemble.randomforestclassifier import RandomForestClassifier
 from cuml.ensemble.randomforestregressor import RandomForestRegressor
