+++ python/cuml/cuml/internals/treelite.pxd
@@ -13,6 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-from cuml.legacy.fil import fil
-from cuml.legacy.fil.fil import ForestInference
+cdef extern from "treelite/c_api.h":
+    ctypedef void* TreeliteModelHandle
+    cdef int TreeliteSerializeModelToBytes(TreeliteModelHandle handle,
+                                           const char** out_bytes, size_t* out_bytes_len)
+    cdef int TreeliteDeserializeModelFromBytes(const char* bytes_seq, size_t len,
+                                               TreeliteModelHandle* out)
+    cdef int TreeliteFreeModel(TreeliteModelHandle handle)
+    cdef const char* TreeliteGetLastError()
