+++ python/cuml/cuml/tests/explainer/test_explainer_kernel_shap.py
@@ -340,7 +340,6 @@ def test_l1_regularization(exact_shap_regression_dataset, l1_type):
     assert isinstance(nz, cp.ndarray)
 
 
-@pytest.mark.skip(reason="Currently failing for unknown reasons.")
 def test_typeerror_input():
     X, y = make_regression(n_samples=100, n_features=10, random_state=10)
     clf = Lasso()
