+++ python/cuml/cuml/tests/test_meta_estimators.py
@@ -118,7 +118,7 @@ def test_pipeline_with_classification(
     assert isinstance(prediction, cupy.ndarray)
     if model_key == "RandomForestClassifier":
         pytest.skip(
-            "RandomForestClassifier is not yet supported"
+            "RandomForestClassifier is not yet supported "
             "by the Pipeline utility"
         )
     _ = pipe.score(X_test, y_test)
