+++ python/cuml/cuml/tests/test_random_forest_compat.py
@@ -52,6 +52,18 @@ from cuml.testing.utils import (
     unit_param,
 )
 
+pytestmark = [
+    pytest.mark.filterwarnings(
+        "ignore: For reproducible results(.*)" "::cuml[.*]"
+    ),
+    pytest.mark.filterwarnings(
+        "ignore: Parameter .* was deprecated in version 25.06.*:FutureWarning"
+    ),
+    pytest.mark.filterwarnings(
+        "ignore: Property .* was deprecated in version 25.06.*:FutureWarning"
+    ),
+]
+
 
 @pytest.fixture(
     scope="session",
@@ -248,7 +260,7 @@ def test_tweedie_convergence(max_depth, split_criterion):
         )
         .fit(X, y)
         .predict(X)
-    ).squeeze()
+    )
     # y should not be non-positive for mean_poisson_deviance
     mask = mse_preds > 0
     mse_tweedie_deviance = mean_tweedie_deviance(
@@ -272,6 +284,9 @@ def test_tweedie_convergence(max_depth, split_criterion):
     cudf_pandas_active,
     reason="cudf.pandas causes sklearn RF estimators crashes sometimes. "
     "Issue: https://github.com/rapidsai/cuml/issues/5991",
+)
+@pytest.mark.filterwarnings(
+    "ignore:Parameter `algo` was deprecated in version 25.06.*:FutureWarning"
 )
 def test_rf_classification(small_clf, datatype, max_samples, max_features):
     use_handle = True
@@ -302,7 +317,11 @@ def test_rf_classification(small_clf, datatype, max_samples, max_features):
     )
     cuml_model.fit(X_train, y_train)
 
-    fil_preds = cuml_model.predict(X_test, predict_model="GPU")
+    # Breaking change for the adoption of new FIL
+    # - algo parameter was deprecated in 25.06 and will be removed in 25.08
+    fil_preds = cuml_model.predict(
+        X_test, predict_model="GPU", threshold=0.5, algo="auto"
+    )
     cu_preds = cuml_model.predict(X_test, predict_model="CPU")
     fil_preds = np.reshape(fil_preds, np.shape(cu_preds))
     cuml_acc = accuracy_score(y_test, cu_preds)
@@ -328,6 +347,9 @@ def test_rf_classification(small_clf, datatype, max_samples, max_features):
     "max_samples", [unit_param(1.0), quality_param(0.90), stress_param(0.95)]
 )
 @pytest.mark.parametrize("datatype", [np.float32, np.float64])
+@pytest.mark.filterwarnings(
+    "ignore:Parameter `algo` was deprecated in version 25.06.*:FutureWarning"
+)
 def test_rf_classification_unorder(
     small_clf, datatype, max_samples, max_features=1, a=2, b=5
 ):
@@ -361,7 +383,11 @@ def test_rf_classification_unorder(
     )
     cuml_model.fit(X_train, y_train)
 
-    fil_preds = cuml_model.predict(X_test, predict_model="GPU")
+    # Breaking change for the adoption of new FIL
+    # - algo parameter was deprecated in 25.06 and will be removed in 25.08
+    fil_preds = cuml_model.predict(
+        X_test, predict_model="GPU", threshold=0.5, algo="auto"
+    )
     cu_preds = cuml_model.predict(X_test, predict_model="CPU")
     fil_preds = np.reshape(fil_preds, np.shape(cu_preds))
     cuml_acc = accuracy_score(y_test, cu_preds)
@@ -713,16 +739,28 @@ def test_rf_classification_proba(
 
 @pytest.mark.parametrize("datatype", [np.float32, np.float64])
 @pytest.mark.parametrize(
-    "fil_layout", ["depth_first", "breadth_first", "layered"]
+    "fil_sparse_format", ["not_supported", True, "auto", False]
+)
+@pytest.mark.parametrize(
+    "algo", ["auto", "naive", "tree_reorg", "batch_tree_reorg"]
 )
 @pytest.mark.skipif(
     cudf_pandas_active,
     reason="cudf.pandas causes sklearn RF estimators crashes sometimes. "
     "Issue: https://github.com/rapidsai/cuml/issues/5991",
 )
-def test_rf_classification_sparse(small_clf, datatype, fil_layout):
+@pytest.mark.filterwarnings(
+    "ignore:Parameter `algo` was deprecated in version 25.06.*:FutureWarning"
+)
+@pytest.mark.filterwarnings(
+    "ignore:Parameter `fil_sparse_format` was deprecated in version 25.06.*"
+    ":FutureWarning"
+)
+def test_rf_classification_sparse(
+    small_clf, datatype, fil_sparse_format, algo
+):
     use_handle = True
-    num_trees = 50
+    num_treees = 50
 
     X, y = small_clf
     X = X.astype(datatype)
@@ -730,7 +768,6 @@ def test_rf_classification_sparse(small_clf, datatype, fil_layout):
     X_train, X_test, y_train, y_test = train_test_split(
         X, y, train_size=0.8, random_state=0
     )
-
     # Create a handle for the cuml model
     handle, stream = get_handle(use_handle, n_streams=1)
 
@@ -742,58 +779,88 @@ def test_rf_classification_sparse(small_clf, datatype, fil_layout):
         min_samples_leaf=2,
         random_state=123,
         n_streams=1,
-        n_estimators=num_trees,
+        n_estimators=num_treees,
         handle=handle,
         max_leaves=-1,
         max_depth=40,
     )
     cuml_model.fit(X_train, y_train)
 
-    fil_preds = cuml_model.predict(
-        X_test,
-        predict_model="GPU",
-        layout=fil_layout,
-    )
-    fil_preds = np.reshape(fil_preds, np.shape(y_test))
-    fil_acc = accuracy_score(y_test, fil_preds)
-    np.testing.assert_almost_equal(fil_acc, cuml_model.score(X_test, y_test))
-
-    fil_model = cuml_model.convert_to_fil_model()
-
-    with cuml.using_output_type("numpy"):
-        fil_model_preds = fil_model.predict(X_test)
-        fil_model_acc = accuracy_score(y_test, fil_model_preds)
-        assert fil_acc == fil_model_acc
-
-    tl_model = cuml_model.convert_to_treelite_model()
-    assert num_trees == tl_model.num_tree
-    assert X.shape[1] == tl_model.num_feature
-
-    if X.shape[0] < 500000:
-        sk_model = skrfc(
-            n_estimators=50,
-            max_depth=40,
-            min_samples_split=2,
-            random_state=10,
+    if (
+        not fil_sparse_format
+        or algo == "tree_reorg"
+        or algo == "batch_tree_reorg"
+    ) or fil_sparse_format == "not_supported":
+        with pytest.raises(ValueError):
+            # Breaking change for the adoption of new FIL
+            # - fil_sparse_format parameter was deprecated in 25.06 and will be
+            #   removed in 25.08
+            # - algo parameter was deprecated in 25.06 and will be removed in
+            #   25.08
+            fil_preds = cuml_model.predict(
+                X_test,
+                predict_model="GPU",
+                threshold=0.5,
+                fil_sparse_format=fil_sparse_format,
+                algo=algo,
+            )
+    else:
+        # Breaking change for the adoption of new FIL
+        # - fil_sparse_format parameter was deprecated in 25.06 and will be
+        #   removed in 25.08
+        # - algo parameter was deprecated in 25.06 and will be removed in 25.08
+        fil_preds = cuml_model.predict(
+            X_test,
+            predict_model="GPU",
+            threshold=0.5,
+            fil_sparse_format=fil_sparse_format,
+            algo=algo,
         )
-        sk_model.fit(X_train, y_train)
-        sk_preds = sk_model.predict(X_test)
-        sk_acc = accuracy_score(y_test, sk_preds)
-        assert fil_acc >= (sk_acc - 0.07)
+        fil_preds = np.reshape(fil_preds, np.shape(y_test))
+        fil_acc = accuracy_score(y_test, fil_preds)
+        np.testing.assert_almost_equal(
+            fil_acc, cuml_model.score(X_test, y_test)
+        )
+
+        fil_model = cuml_model.convert_to_fil_model()
+
+        with cuml.using_output_type("numpy"):
+            fil_model_preds = fil_model.predict(X_test)
+            fil_model_acc = accuracy_score(y_test, fil_model_preds)
+            assert fil_acc == fil_model_acc
+
+        tl_model = cuml_model.convert_to_treelite_model()
+        assert num_treees == tl_model.num_tree  # breaking change in 25.06
+        assert X.shape[1] == tl_model.num_feature  # breaking change in 25.06
+
+        if X.shape[0] < 500000:
+            sk_model = skrfc(
+                n_estimators=50,
+                max_depth=40,
+                min_samples_split=2,
+                random_state=10,
+            )
+            sk_model.fit(X_train, y_train)
+            sk_preds = sk_model.predict(X_test)
+            sk_acc = accuracy_score(y_test, sk_preds)
+            assert fil_acc >= (sk_acc - 0.07)
 
 
 @pytest.mark.parametrize("datatype", [np.float32, np.float64])
 @pytest.mark.parametrize(
-    "fil_layout", ["depth_first", "breadth_first", "layered"]
+    "fil_sparse_format", ["not_supported", True, "auto", False]
+)
+@pytest.mark.parametrize(
+    "algo", ["auto", "naive", "tree_reorg", "batch_tree_reorg"]
 )
 @pytest.mark.skipif(
     cudf_pandas_active,
     reason="cudf.pandas causes sklearn RF estimators crashes sometimes. "
     "Issue: https://github.com/rapidsai/cuml/issues/5991",
 )
-def test_rf_regression_sparse(special_reg, datatype, fil_layout):
+def test_rf_regression_sparse(special_reg, datatype, fil_sparse_format, algo):
     use_handle = True
-    num_trees = 50
+    num_treees = 50
 
     X, y = special_reg
     X = X.astype(datatype)
@@ -812,7 +879,7 @@ def test_rf_regression_sparse(special_reg, datatype, fil_layout):
         min_samples_leaf=2,
         random_state=123,
         n_streams=1,
-        n_estimators=num_trees,
+        n_estimators=num_treees,
         handle=handle,
         max_leaves=-1,
         max_depth=40,
@@ -821,51 +888,63 @@ def test_rf_regression_sparse(special_reg, datatype, fil_layout):
     cuml_model.fit(X_train, y_train)
 
     # predict using FIL
-    fil_preds = cuml_model.predict(
-        X_test,
-        predict_model="GPU",
-        layout=fil_layout,
-    )
-    fil_preds = np.reshape(fil_preds, np.shape(y_test))
-    fil_r2 = r2_score(y_test, fil_preds)
-
-    fil_model = cuml_model.convert_to_fil_model()
-
-    with cuml.using_output_type("numpy"):
-        fil_model_preds = fil_model.predict(X_test)
-        fil_model_preds = np.reshape(fil_model_preds, np.shape(y_test))
-        fil_model_r2 = r2_score(y_test, fil_model_preds)
-        assert fil_r2 == fil_model_r2
-
-    tl_model = cuml_model.convert_to_treelite_model()
-    assert num_trees == tl_model.num_tree
-    assert X.shape[1] == tl_model.num_feature
-
-    # Initialize, fit and predict using
-    # sklearn's random forest regression model
-    if X.shape[0] < 1000:  # mode != "stress":
-        sk_model = skrfr(
-            n_estimators=50,
-            max_depth=40,
-            min_samples_split=2,
-            random_state=10,
+    if (
+        not fil_sparse_format
+        or algo == "tree_reorg"
+        or algo == "batch_tree_reorg"
+    ) or fil_sparse_format == "not_supported":
+        with pytest.raises(ValueError):
+            fil_preds = cuml_model.predict(
+                X_test,
+                predict_model="GPU",
+                fil_sparse_format=fil_sparse_format,
+                algo=algo,
+            )
+    else:
+        fil_preds = cuml_model.predict(
+            X_test,
+            predict_model="GPU",
+            fil_sparse_format=fil_sparse_format,
+            algo=algo,
         )
-        sk_model.fit(X_train, y_train)
-        sk_preds = sk_model.predict(X_test)
-        sk_r2 = r2_score(y_test, sk_preds)
-        assert fil_r2 >= (sk_r2 - 0.08)
+        fil_preds = np.reshape(fil_preds, np.shape(y_test))
+        fil_r2 = r2_score(y_test, fil_preds)
+
+        fil_model = cuml_model.convert_to_fil_model()
+
+        with cuml.using_output_type("numpy"):
+            fil_model_preds = fil_model.predict(X_test)
+            fil_model_preds = np.reshape(fil_model_preds, np.shape(y_test))
+            fil_model_r2 = r2_score(y_test, fil_model_preds)
+            assert fil_r2 == fil_model_r2
+
+        tl_model = cuml_model.convert_to_treelite_model()
+        assert num_treees == tl_model.num_tree  # breaking change in 25.06
+        assert X.shape[1] == tl_model.num_feature  # breaking change in 25.06
+
+        # Initialize, fit and predict using
+        # sklearn's random forest regression model
+        if X.shape[0] < 1000:  # mode != "stress":
+            sk_model = skrfr(
+                n_estimators=50,
+                max_depth=40,
+                min_samples_split=2,
+                random_state=10,
+            )
+            sk_model.fit(X_train, y_train)
+            sk_preds = sk_model.predict(X_test)
+            sk_r2 = r2_score(y_test, sk_preds)
+            assert fil_r2 >= (sk_r2 - 0.08)
 
 
 @pytest.mark.xfail(reason="Need rapidsai/rmm#415 to detect memleak robustly")
 @pytest.mark.memleak
 @pytest.mark.parametrize("datatype", [np.float32, np.float64])
-@pytest.mark.parametrize(
-    "fil_layout", ["depth_first", "breadth_first", "layered"]
-)
+@pytest.mark.parametrize("fil_sparse_format", [True, False, "auto"])
 @pytest.mark.parametrize(
     "n_iter", [unit_param(5), quality_param(30), stress_param(80)]
 )
-def test_rf_memory_leakage(small_clf, datatype, fil_layout, n_iter):
+def test_rf_memory_leakage(small_clf, datatype, fil_sparse_format, n_iter):
     use_handle = True
 
     X, y = small_clf
@@ -874,7 +953,6 @@ def test_rf_memory_leakage(small_clf, datatype, fil_layout, n_iter):
     X_train, X_test, y_train, y_test = train_test_split(
         X, y, train_size=0.8, random_state=0
     )
-
     # Create a handle for the cuml model
     handle, stream = get_handle(use_handle, n_streams=1)
 
@@ -887,7 +965,6 @@ def test_rf_memory_leakage(small_clf, datatype, fil_layout, n_iter):
     free_mem = cuda.current_context().get_memory_info()[0]
 
     def test_for_memory_leak():
-        nonlocal free_mem
         cuml_mods = curfc(handle=handle)
         cuml_mods.fit(X_train, y_train)
         handle.sync()  # just to be sure
@@ -899,7 +976,7 @@ def test_rf_memory_leakage(small_clf, datatype, fil_layout, n_iter):
             cuml_mods.predict(
                 X_test,
                 predict_model="GPU",
-                layout=fil_layout,
+                fil_sparse_format=fil_sparse_format,
             )
             handle.sync()  # just to be sure
             # Calculate the memory free after predicting the cuML model
@@ -1140,12 +1217,11 @@ def test_rf_get_json(estimator_type, max_depth, n_estimators):
             majority_vote = predict_with_json_rf_classifier(json_obj, row)
             assert expected_pred[idx] == majority_vote
     elif estimator_type == "regression":
-        expected_pred = cuml_model.predict(X).astype(np.float32).squeeze()
+        expected_pred = cuml_model.predict(X).astype(np.float32)
         pred = []
         for idx, row in enumerate(X):
             pred.append(predict_with_json_rf_regressor(json_obj, row))
         pred = np.array(pred, dtype=np.float32)
-        print(json_obj)
         for i in range(len(pred)):
             assert np.isclose(pred[i], expected_pred[i]), X[i, 19]
         np.testing.assert_almost_equal(pred, expected_pred, decimal=6)
@@ -1325,6 +1401,7 @@ def test_rf_regressor_gtil_integration(tmpdir):
     expected_pred = clf.predict(X).reshape((-1, 1, 1))
 
     checkpoint_path = os.path.join(tmpdir, "checkpoint.tl")
+    # breaking change in 25.06: use serialize() instead of to_treelite_checkpoint()
     clf.convert_to_treelite_model().serialize(checkpoint_path)
 
     tl_model = treelite.Model.deserialize(checkpoint_path)
@@ -1340,6 +1417,7 @@ def test_rf_binary_classifier_gtil_integration(tmpdir):
     expected_pred = clf.predict_proba(X).reshape((-1, 1, 2))
 
     checkpoint_path = os.path.join(tmpdir, "checkpoint.tl")
+    # breaking change in 25.06: use serialize() instead of to_treelite_checkpoint()
     clf.convert_to_treelite_model().serialize(checkpoint_path)
 
     tl_model = treelite.Model.deserialize(checkpoint_path)
@@ -1355,6 +1433,7 @@ def test_rf_multiclass_classifier_gtil_integration(tmpdir):
     expected_prob = clf.predict_proba(X).reshape((X.shape[0], 1, -1))
 
     checkpoint_path = os.path.join(tmpdir, "checkpoint.tl")
+    # breaking change in 25.06: use serialize() instead of to_treelite_checkpoint()
     clf.convert_to_treelite_model().serialize(checkpoint_path)
 
     tl_model = treelite.Model.deserialize(checkpoint_path)
@@ -1406,7 +1485,6 @@ def test_random_forest_max_features_deprecation(Estimator):
 def test_rf_predict_returns_int():
 
     X, y = make_classification()
-
     # Capture and verify expected warning
     warning_msg = (
         "The number of bins, `n_bins` is greater than the number of samples "
